refactor: split oversized files + enforce a 500-line CI cap
Five files had accumulated multiple responsibilities. Split them along
their natural seams (behaviour-preserving moves, no logic changes), then
add a hard file-length cap so they can't silently grow back.

Logic extractions (now unit-testable in isolation):
- HomeViewModel (517→334): AI-thread/content-block parsing → AiThreadMapper.
- AiTurnWorker (344→218): prompt construction → AiPromptBuilder.

Compose extractions to components/ (internal):
- AiThinkingThread (921 → AiThinkingThread/ThinkingTimeline/ProcessRows/
  ResponseMarkdown/MarkdownTable).
- SettingsScreen (673→268): Section, ActionRow, TimePickerRow, BufferSlider,
  TextEntryRows, DailyBudgetRow → settings/components/.
- HomeScreen (559→315): SettingsChangedPill, OnboardingHint,
  NotificationPermissionRow, AiFailureBanner → home/components/.

Left cohesive files alone: NextAlarmCard (422), OnboardingScreen (388),
CronBottomBar (301).

CI cap:
- app/build.gradle.kts: `checkFileLength` task fails if any src/main Kotlin
  file exceeds 500 lines; wired into `check` and the build workflow.
- AGENTS.md: documents the cap as a regression backstop.

Tests: AiThreadMapperTest, AiPromptBuilderTest. Verified with assembleDebug,
lintDebug, testDebugUnitTest, checkFileLength; visually confirmed Home +
Settings on an emulator.
