+++ AGENTS.md
@@ -56,6 +56,7 @@ These rules exist because LLM passes have repeatedly violated them. Follow them 
 - Default to writing no comment. Add one only when the *why* is non-obvious (a workaround for a specific bug, a hidden invariant, a constraint from an external system). Don't restate what the code does — clean names and whitespace explain far more than prose.
 - No back-to-back `//` paragraph comments above a declaration. Several `//` lines forming a paragraph should become a single `/** … */` KDoc placed **immediately above** the declaration: KDoc flows into IDE quick-docs and hover; stacked `//` lines do not. (A single inline `//` explaining one line *inside* a function body is fine.)
 - No section-banner comments (`// ──── Foo ────`, `// ---- Foo ----`, `/* ===== Foo ===== */`). They're a code smell that signals a file is doing too much. **Split into atomic files instead** — one file, one responsibility — rather than partitioning one file with banners.
+- **Hard cap: 500 lines per `src/main` Kotlin file**, enforced by the `checkFileLength` Gradle task in CI (`./gradlew :app:checkFileLength`). The cap is a regression backstop, not the rule — the rule is one file, one responsibility, and most files should sit far below it (the repo mean is ~130). When a file approaches the cap, split it (extract reusable composables to a `components/` package, pull pure parsing/mapping logic out of a ViewModel/Worker into a testable object); don't suppress. A genuinely cohesive single unit near the cap (e.g. one self-contained card composable) is fine.
 - Don't add a multi-paragraph KDoc block to a private composable describing what it renders — the name and signature already say that. One short line max.
 - Don't leave `// TODO`, `// for now`, or hand-tuned magic numbers without a measurement-based replacement plan. The pattern of `Modifier.offset(x = (-6).dp) // compensate for X` has been removed from this repo (see `AlignedFirstGlyph` in `NextAlarmCard.kt` for the proper measurement-based approach); don't reintroduce it.
