+++ app/src/main/java/fr/bsodium/cron/ui/screens/settings/components/ActionRow.kt
@@ -1,0 +1,56 @@
+package fr.bsodium.cron.ui.screens.settings.components
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import fr.bsodium.cron.ui.theme.Spacing
+
+/** A title/subtitle row that shows either an "Enabled" badge or an action button. */
+@Composable
+internal fun ActionRow(
+    title: String,
+    subtitle: String,
+    done: Boolean,
+    actionLabel: String,
+    enabled: Boolean,
+    onClick: () -> Unit,
+) {
+    Row(
+        modifier = Modifier.fillMaxWidth(),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        Column(modifier = Modifier.weight(1f)) {
+            Text(
+                text = title,
+                style = MaterialTheme.typography.bodyLarge,
+                color = MaterialTheme.colorScheme.onBackground,
+            )
+            Text(
+                text = subtitle,
+                style = MaterialTheme.typography.bodySmall,
+                color = MaterialTheme.colorScheme.onSurfaceVariant,
+            )
+        }
+        if (done) {
+            Text(
+                text = "Enabled",
+                style = MaterialTheme.typography.titleMedium,
+                color = MaterialTheme.colorScheme.primary,
+                fontWeight = FontWeight.SemiBold,
+                modifier = Modifier.padding(horizontal = Spacing.md, vertical = Spacing.sm),
+            )
+        } else {
+            TextButton(onClick = onClick, enabled = enabled) {
+                Text(actionLabel, color = MaterialTheme.colorScheme.primary)
+            }
+        }
+    }
+}
