+++ app/src/main/java/fr/bsodium/cron/worker/AiTurnWorker.kt
@@ -34,19 +34,13 @@ import fr.bsodium.cron.calendar.CalendarReader
 import fr.bsodium.cron.session.SessionRepository
 import fr.bsodium.cron.session.db.CronDatabase
 import fr.bsodium.cron.session.model.ActionType
-import fr.bsodium.cron.session.model.EventData
-import fr.bsodium.cron.session.model.SessionEvent
 import fr.bsodium.cron.session.model.SleepSession
-import fr.bsodium.cron.session.model.SignalConfidence
 import fr.bsodium.cron.session.model.TriggerType
 import fr.bsodium.cron.settings.SecureKeyStore
 import fr.bsodium.cron.settings.SettingsRepository
 import fr.bsodium.cron.travel.GeocodingClient
 import fr.bsodium.cron.travel.RoutesClient
-import kotlinx.datetime.Clock
 import kotlinx.datetime.TimeZone
-import kotlinx.datetime.toLocalDateTime
-import kotlin.time.Duration.Companion.minutes
 
 /**
  * Resumable WorkManager worker that drives one AI tool-use turn for a session.
@@ -115,7 +109,8 @@ class AiTurnWorker(
             thinking = thinking,
         )
 
-        val userMessage = buildUserMessage(session, isEveningPlan)
+        val instructions = settingsRepository.currentUserInstructions()
+        val userMessage = AiPromptBuilder.build(session, isEveningPlan, instructions)
 
         return try {
             val outcome = runner.run(sessionId, turnIndex, userMessage)
@@ -171,114 +166,6 @@ class AiTurnWorker(
         return ToolRegistry(tools)
     }
 
-    private suspend fun buildUserMessage(session: SleepSession, isEveningPlan: Boolean): String {
-        val now = Clock.System.now()
-        val tz = TimeZone.of(session.timezone)
-        val localNow = now.toLocalDateTime(tz)
-        val instructions = settingsRepository.currentUserInstructions()
-
-        return if (isEveningPlan) buildEveningPlanMessage(session, localNow, tz, instructions)
-        else buildOvernightReplanMessage(session, localNow, instructions)
-    }
-
-    private fun buildEveningPlanMessage(
-        session: SleepSession,
-        localNow: kotlinx.datetime.LocalDateTime,
-        tz: TimeZone,
-        userInstructions: String?,
-    ): String {
-        val plan = session.plan
-        // Latest evening-plan event, so a manual replan's freshly-captured location wins.
-        val eveningEvent = session.events.lastOrNull { it.trigger == TriggerType.EveningPlan }
-        val location = (eveningEvent?.data as? EventData.EveningPlan)?.location
-
-        return buildString {
-            appendLine("It is now $localNow (${session.timezone}).")
-            appendLine()
-            appendLine("## Session context")
-            appendLine("- Morning date: ${session.date}")
-            appendLine("- Hard latest (never exceed): ${plan.hardLatest}")
-            appendLine("- Wake window: ${plan.wakeWindowStart} – ${plan.wakeWindowEnd}")
-            appendLine("- Travel buffer (minimum commute floor, NOT prep): ${plan.commuteBufferMinutes} min")
-            appendLine("- Morning preparation time (getting ready, separate from travel): ${plan.preparationBufferMinutes} min")
-            appendLine("- Free day wake window: ${plan.wakeWindowStart} – ${plan.wakeWindowEnd}")
-            appendLine()
-            if (location != null) {
-                appendLine("## Current location")
-                appendLine("- Latitude: ${location.lat}, Longitude: ${location.lng}")
-                appendLine("- Source: ${location.source.name.lowercase()}")
-                appendLine("- Accuracy: ±${location.accuracyMeters?.let { "${it.toInt()} m" } ?: "unknown"}")
-                appendLine("- Captured at: ${location.capturedAt}")
-            } else {
-                appendLine("## Location")
-                appendLine("- Source: unavailable — apply a flat +30 min buffer and mention it in the reason")
-            }
-            appendLine()
-            appendUserInstructions(userInstructions)
-            appendLine("Plan tomorrow's alarm. Follow the process in your system prompt: read calendar, identify anchor event, estimate commute if applicable, then call set_alarm.")
-        }
-    }
-
-    /** Appends the user's standing custom instructions, when set, as a labelled block. */
-    private fun StringBuilder.appendUserInstructions(text: String?) {
-        if (text.isNullOrBlank()) return
-        appendLine("## User instructions")
-        appendLine("Standing instructions from the user — honour them unless they'd push the alarm past the hard latest:")
-        appendLine(text)
-        appendLine()
-    }
-
-    private fun isPhoneOnlyMode(session: SleepSession): Boolean {
-        val sessionAge = Clock.System.now() - session.createdAt
-        if (sessionAge < PHONE_ONLY_THRESHOLD) return false
-        return session.events.none {
-            it.trigger == TriggerType.HcStageUpdate &&
-                (it.data as? EventData.HcStageUpdate)?.confidence == SignalConfidence.High
-        }
-    }
-
-    private fun buildOvernightReplanMessage(
-        session: SleepSession,
-        localNow: kotlinx.datetime.LocalDateTime,
-        userInstructions: String?,
-    ): String {
-        val plan = session.plan
-        val instr = session.currentInstruction
-
-        return buildString {
-            appendLine("It is now $localNow (${session.timezone}).")
-            appendLine()
-            if (isPhoneOnlyMode(session)) {
-                appendLine("**Note: phone-only mode is active.** No high-confidence Health Connect data has arrived in the past 90 minutes. Rely exclusively on screen-state and activity signals. Do not request or expect sleep stage updates; treat any stage signals as low-confidence approximations.")
-                appendLine()
-            }
-            appendLine("## Day plan")
-            appendLine("- Morning date: ${session.date}")
-            appendLine("- Hard latest (never exceed): ${plan.hardLatest}")
-            appendLine("- Wake window: ${plan.wakeWindowStart} – ${plan.wakeWindowEnd}")
-            appendLine("- Snooze count so far: ${session.snoozeCount}")
-            appendLine()
-            appendLine("## Current instruction")
-            appendLine("- Action: ${instr.action.name}")
-            if (instr.alarmTime != null) appendLine("- Alarm set for: ${instr.alarmTime}")
-            appendLine("- Reason: ${instr.reason}")
-            appendLine()
-            appendLine("## Event log (oldest first)")
-            session.events.forEachIndexed { i, ev ->
-                appendLine("${i + 1}. [${ev.timestamp}] ${ev.trigger.name}: ${summarizeEventData(ev)}")
-            }
-            appendLine()
-            val last = session.events.lastOrNull()
-            if (last != null) {
-                appendLine("## Triggering event")
-                appendLine("[${last.timestamp}] ${last.trigger.name}: ${summarizeEventData(last)}")
-            }
-            appendLine()
-            appendUserInstructions(userInstructions)
-            appendLine("Decide what the alarm system should do. Call set_alarm if you want to adjust the wake time. If the current alarm is already optimal, respond with a brief explanation and do not call any tool.")
-        }
-    }
-
     private suspend fun postPlanningNotification(sessionId: String) {
         val updated = repository.findById(sessionId) ?: return
         val instruction = updated.currentInstruction
@@ -313,18 +200,6 @@ class AiTurnWorker(
         }
     }
 
-    private fun summarizeEventData(event: SessionEvent): String = when (val d = event.data) {
-        is EventData.EveningPlan -> "timezone=${d.timezone}, location_source=${d.location.source}"
-        is EventData.SleepOnset -> "screen_off_since=${d.screenOffSince}, rearm=${d.rearm}"
-        is EventData.HcStageUpdate -> "stage=${d.stage}, source=${d.source}, confidence=${d.confidence}"
-        is EventData.MidSleepActivity -> "activity=${d.activityType}, screen_on=${d.screenOn}, duration=${d.durationSeconds}s"
-        is EventData.OutOfBedConfirmed -> "evidence=${d.evidence}"
-        is EventData.WakeWindowOpportunity -> "stage=${d.currentStage}, window=${d.windowStart}–${d.windowEnd}"
-        is EventData.AlarmInteraction -> "snooze_duration=${d.snoozeDurationMinutes}min, count=${d.snoozeCount}"
-        is EventData.CalendarChange -> "type=${d.changeType}, affects_first=${d.affectsFirstEvent}"
-        EventData.Empty -> "(empty)"
-    }
-
     companion object {
         const val KEY_SESSION_ID = "session_id"
         const val WORK_PREFIX = "ai_turn_"
@@ -338,7 +213,6 @@ class AiTurnWorker(
         const val REASON_HTTP = "http_error"
 
         private const val TAG = "AiTurnWorker"
-        private val PHONE_ONLY_THRESHOLD = 90.minutes
         private const val THINKING_BUDGET = 2_000
     }
 }
